+++ common/database/model/metadata/metadata.go
@@ -44,28 +44,6 @@ type Token struct {
 	NFTMetadata json.RawMessage `json:"nft_metadata,omitempty"`
 }
 
-type Mirror struct {
-	ContentType           string        `json:"content_type"`
-	Contributor           string        `json:"contributor"`
-	ContentDigest         string        `json:"content_digest"`
-	OriginalContentDigest string        `json:"original_content_digest,omitempty"`
-	Content               MirrorContent `json:"content"`
-}
-
-type MirrorContent struct {
-	Nft     json.RawMessage `json:"nft"`
-	Wnft    json.RawMessage `json:"wnft"`
-	Digest  string          `json:"digest"`
-	Content struct {
-		Body      string          `json:"body"`
-		Title     string          `json:"title"`
-		Timestamp decimal.Decimal `json:"timestamp"`
-	} `json:"content"`
-	Version        string          `json:"version"`
-	Authorship     json.RawMessage `json:"authorship"`
-	OriginalDigest string          `json:"originalDigest"`
-}
-
 const (
 	SwapTypeRouter = "router"
 	SwapTypePool   = "pool"
